+++ server/src/anydns/mod.rs
@@ -10,6 +10,7 @@ mod dns_socket;
 mod pending_request;
 mod query_id_manager;
 mod server;
+mod rate_limiter;
 
 pub use custom_handler::{CustomHandler, CustomHandlerError, EmptyHandler, HandlerHolder};
 pub use dns_socket::DnsSocket;
